FUSETOOLS2-789 - move util test class for Camel Kafka Connector Catalog in a more common place
